+++ client/src/pages/Session.jsx
@@ -47,9 +47,8 @@ const Session = () => {
   return (
     <main className='stndrd-page'>
       {/* <div className='container'> */}
-      <h4>{user.username}s Imaging Journal</h4>
+      <h4>{user.username}'s Session Logs</h4>
       <div className='row'>
-        <h5>Sessions</h5>
         <div className='row'>
           <div className='col S12 m3 l2'> Filters</div>
           <div className='col S12 m9 l10'>